Allow clean first time setup.
This commit fix a problem related to the first installation.
Previous commits assumed that this package was installed properly
at least once. Consequently it assumed
that the LLVS_DEFAULT_ILDS variable was set correctly.
This variable is now set directly inside the main CMakeLists.txt
file, and is propagated across the tree. Most notably the file
LLVT-idlConfig.cmake is now generated automatically.